homework v1

+++ Lab3/src/CargoCapable.java
@@ -1,4 +1,4 @@
 public interface CargoCapable {
-    int getMaxPayload();
+    double getMaxPayload();
 
 }
