+++ Lab3/src/Freighter.java
@@ -1,12 +1,12 @@
 public class Freighter extends Aircraft implements CargoCapable{
-    private int maxPayload;
+    private double maxPayload;
 
     public Freighter(String name, String model, String tailNumber, int maxPayload) {
         super(name, model, tailNumber);
         this.maxPayload = maxPayload;
     }
 
-    public int getMaxPayload() {
+    public double getMaxPayload() {
         return maxPayload;
     }
 }
